rename item to baseitem, add wrapper item to check on props
