+++ packages/react/src/ActionList/Item.test.tsx
@@ -13,9 +13,9 @@ function SimpleActionList(): JSX.Element {
       <ActionList.Item>Copy link</ActionList.Item>
       <ActionList.Item>Edit file</ActionList.Item>
       <ActionList.Item variant="danger">Delete file</ActionList.Item>
-      <ActionList.LinkItem href="//github.com" title="anchor" aria-keyshortcuts="d">
+      <ActionList.Item href="//github.com" title="anchor" aria-keyshortcuts="d">
         Link Item
-      </ActionList.LinkItem>
+      </ActionList.Item>
       <ActionList.Item inactiveText="Unavailable due to an outage">Inactive item</ActionList.Item>
       <ActionList.Item inactiveText="Unavailable due to an outage" loading>
         Loading and inactive item
@@ -63,6 +63,7 @@ describe('ActionList.Item', () => {
   it('should have aria-keyshortcuts applied to the correct element', async () => {
     const {container} = HTMLRender(<SimpleActionList />)
     const linkOptions = await waitFor(() => container.querySelectorAll('a'))
+    console.log(linkOptions[0])
     expect(linkOptions[0]).toHaveAttribute('aria-keyshortcuts', 'd')
     expect(linkOptions[0].parentElement).not.toHaveAttribute('aria-keyshortcuts', 'd')
   })
@@ -185,9 +186,9 @@ describe('ActionList.Item', () => {
     const onClick = vi.fn()
     const component = HTMLRender(
       <ActionList role="listbox">
-        <ActionList.LinkItem role="link" onClick={onClick}>
+        <ActionList.Item role="link" onClick={onClick}>
           Primer React
-        </ActionList.LinkItem>
+        </ActionList.Item>
       </ActionList>,
     )
     const link = await waitFor(() => component.getByRole('link'))
