+++ packages/react/src/Pagination/mocks/ReactRouterLink.tsx
@@ -5,7 +5,7 @@ export type ReactRouterLikeLinkProps = {to: string; children: React.ReactNode; c
 export const ReactRouterLikeLink = React.forwardRef<HTMLAnchorElement, ReactRouterLikeLinkProps>(
   ({to, children, ...props}, ref) => {
     return (
-      <a ref={ref} href={to} {...props}>
+      <a ref={ref} {...props} href={to}>
         {children}
       </a>
     )
